moved BuildWorkfile to avalon context
